Fixes vs2013 warnings 4251, 4267.
The pragma to ignore 4251 already existed in windowsIncludes.h, and it works
when compiling lepton itself. However, when clients use lepton, they get this
warning. I think this is why @sherm1 put the define as shown here:
https://github.com/simbody/simbody/blob/master/SimTKcommon/include/SimTKcommon/internal/common.h#L168
 (whether or not the shared library is being built or used).

+++ libraries/lepton/src/CompiledExpression.cpp
@@ -121,7 +121,7 @@ void CompiledExpression::compileExpression(const ExpressionTreeNode& node, vecto
                 arguments[stepIndex] = args;
         }
     }
-    temps.push_back(make_pair(node, workspace.size()));
+    temps.push_back(make_pair(node, (int) workspace.size()));
     workspace.push_back(0.0);
 }
